requirements.txt
openpyxl==2.6.2
requests==2.28.2
progressbar2==4.2.0 